Creates search store
